refactor(runtime): split util.rs god object into modules
