+++ sandbox-runtime/src/util/shell.rs
@@ -1,0 +1,4 @@
+pub fn shell_escape(value: &str) -> String {
+    let escaped = value.replace('\'', "'\"'\"'");
+    format!("'{escaped}'")
+}
