+++ scripts/.god-object-baseline.txt
@@ -5,4 +5,3 @@
 738 ui/src/routes/sandboxes.$id.tsx
 653 ai-agent-instance-blueprint-lib/src/auto_provision.rs
 634 ui/src/routes/instances.$id.tsx
-606 sandbox-runtime/src/util.rs
